feat: change API for inclusion list node to not include metadata (#162)
Changing the API we use to fetch inclusion lists to strip out the extra metadata we had there before. Now we just decode the RLP bytes (cheap), extract the metadata we want to store as debug info in postgres, & send on to the simulator for validation.

+++ crates/common/src/utils.rs
@@ -163,7 +163,7 @@ pub fn utcnow_ns() -> u64 {
 
 #[inline(always)]
 pub fn get_slot_coordinate(
-    slot: i32,
+    slot: u64,
     proposer_pub_key: &BlsPublicKey,
     parent_hash: &B256,
 ) -> String {
